backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
    "django>=4.1.13",
    "django-cors-headers>=4.5.0",
    "django-storages[s3]>=1.14.6",
    "djangorestframework>=3.15.1",
    "djangorestframework-simplejwt>=5.3.1",
    "djongo>=1.3.6",
    "dnspython>=2.7.0",
    "gunicorn>=23.0.0",
    "pillow>=10.4.0",
    "pymongo<4.0",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "whitenoise>=6.11.0",
]
